Crate ogmios_client

This module provides a high-level API for interacting with the Ogmios JSON-RPC API.

Ogmios is a JSON-RPC server that provides a high-level API for interacting with the Cardano blockchain. It can be accessed via a HTTP or WebSocket connection.

More information about Ogmios API can be found at https://ogmios.dev/api/

Modules

jsonrpsee

OgmiosClient implementation with jsonrpsee. Major drawback is that it swallows the error response from the server in case of 400 Bad Request.

query_ledger_state

Queries that start with queryLedgerState/.

query_network

Queries that start with queryNetwork/.

transactions

Requests to evalute and submit transactions via Ogmios`.

types

Common types used in the Ogmios API.

Structs

ByNameParamsBuilder

Builder for named parameters.

Enums

OgmiosClientError

Represents an error that can occur when interacting with the Ogmios JSON-RPC API.

OgmiosParams

Enum representing the parameters for a JSON-RPC request to the Ogmios server.

Traits

OgmiosClient

Trait for interacting with the Ogmios JSON-RPC API.