ogmios_client

Module query_ledger_state

Queries that start with queryLedgerState/.

Structs

EpochBoundary

Represents the boundary of an epoch.

EpochParameters

Represents the parameters of an epoch.

EraSummary

Represents a summary of a known era.

OgmiosTip

Represents the tip of the ledger.

PlutusCostModels

Represents the cost models for different Plutus language versions.

ProtocolParametersResponse

Represents the protocol parameters.

ReferenceScriptsCosts

Represents the cost of reference scripts.

ScriptExecutionPrices

Represents the pricing for Plutus script execution resources.

Traits

QueryLedgerState

Trait that defines the methods for querying the Cardano ledger state.

QueryUtxoByUtxoId

Trait that defines the methods for querying a single UTXO by transaction hash and output index.